Fix the bot
