+++ nfcli/bot.py
@@ -2,19 +2,28 @@ import logging
 import os
 import re
 from posixpath import basename
+from typing import List
 
+import discord
 from discord import File, Message
 
-from nfcli import load_path
+from nfcli import DISCORD_TOKEN, init_logger, load_path
 from nfcli.parser import parse_any, parse_mods
 from nfcli.printer import FleetPrinter
 from nfcli.steam import get_workshop_files, get_workshop_id
+from nfcli.wiki import Wiki
 from nfcli.writer import determine_output_png, get_temp_filename
+
+wiki_db = Wiki()
+intents = discord.Intents.default()
+intents.message_content = True
+bot = discord.Bot(intents=intents)
+init_logger("bot.log", logging.INFO)
 
 
 async def process_file(message: Message, xml_data: str, filename: str, with_fleet_file: bool):
     """Process one file content."""
-    logging.info(f"Converting {filename}")
+    logging.info(f"Converting file {filename}")
     png_file = determine_output_png(filename)
     tmp_file = get_temp_filename(".png")
     entity = parse_any(filename, xml_data)
@@ -34,9 +43,11 @@ async def process_uploads(message: Message):
     """Process uploaded files."""
     ship_files = [file for file in message.attachments if file.filename.endswith("ship")]
     fleet_files = [file for file in message.attachments if file.filename.endswith("fleet")]
-    for file in ship_files + fleet_files:
-        xml_data = await file.read()
-        await process_file(message, xml_data, file.filename, with_fleet_file=False)
+    all_files = ship_files + fleet_files
+    for file in all_files:
+        async with message.channel.typing():
+            xml_data = await file.read()
+            await process_file(message, xml_data, file.filename, with_fleet_file=False)
 
 
 async def process_workshop(message: Message, workshop_id: int):
@@ -61,4 +72,45 @@ async def process_workshops(message: Message):
         workshop_ids.add(get_workshop_id(link_data))
 
     for workshop_id in workshop_ids:
-        await process_workshop(message, workshop_id)
+        async with message.channel.typing():
+            await process_workshop(message, workshop_id)
+
+
+async def replace_with_previous(channel: discord.TextChannel, link: str, message: str) -> str:
+    old_messages: List[discord.Message] = await channel.history(limit=100).flatten()
+    for old_message in old_messages:
+        if old_message.author != bot.user:
+            continue
+        if link in old_message.content:
+            return f"I have already explained this recently!\n<{old_message.jump_url}>"
+    return message
+
+
+@bot.event
+async def on_ready():
+    logging.info("Workshop bot initialized")
+    for guild in bot.guilds:
+        logging.info(f"Connected to the guild: {guild.name} (id: {guild.id})")
+
+
+@bot.event
+async def on_message(message: discord.Message):
+    if message.author == bot.user:
+        return
+    await process_workshops(message),
+    await process_uploads(message),
+
+
+@bot.command(name="wiki")
+async def wiki(ctx: discord.ApplicationContext, *, keywords):
+    """Search N:FC wiki data dumps (thanks to @Alexbay218#0295)"""
+    async with ctx.typing():
+        entity = wiki_db.get(keywords)
+        message = entity.text
+        if entity:
+            message = await replace_with_previous(ctx.channel, entity.link, message)
+        await ctx.respond(message)
+
+
+def start():
+    bot.run(DISCORD_TOKEN)
